move Rich Text Functions new module
